% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accuracyPrecision.R
\name{accuracyPrecision}
\alias{accuracyPrecision}
\title{Plot targets and histograms for illustrating accuracy and precision.}
\usage{
accuracyPrecision(n = 50, r = 4, pts.col = "red", pts.trans = 1,
  pts.cex = 1, mns.col = "blue", mns.pch = 3, mns.lwd = 2)
}
\arguments{
\item{n}{Number of random points to put on each target and histogram.}

\item{r}{Number of rings (i.e., circles) on the target.}

\item{pts.col}{String indicating color to use for plotting points on the
target.}

\item{pts.trans}{Value between 0 and 1 indicating the transparency to use for
plotting points on the target.  The inverse of this value indicates how many
points must be overplotted before the full color is shown.}

\item{pts.cex}{Character expansion amount for points on target.}

\item{mns.col}{String indicating color to use for line marking the mean on
the targets and the histograms.}

\item{mns.pch}{Value indicating the plotting character to use for the mean on
the targets.}

\item{mns.lwd}{Value indicating the line width to use use for the mean on the
histograms.}
}
\value{
None.  A plot is produced.
}
\description{
Plot \dQuote{targets} and histograms for illustrating accuracy and precision.
The user can choose the number of points to sample.
}
\examples{
accuracyPrecision()
accuracyPrecision(100,5)
}
\keyword{dynamic}

