% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptionCheck.R
\name{assumptionCheck}
\alias{assumptionCheck}
\alias{transChooser}
\title{Graphics for assessing assumptions of simple linear models.}
\usage{
assumptionCheck(
  object,
  lambday = 1,
  lambdax = 1,
  shifty = 0,
  shiftx = 0,
  show.stats = TRUE,
  boxplot = TRUE,
  alpha = 0.05,
  col.hist = "gray90"
)
}
\arguments{
\item{object}{An \code{lm} object or formula depicting a one-way or two-way ANOVA model or a simple linear or indicator variable regression model.}

\item{lambday}{A numeric value for the power of the transformation of the response variable (see details).}

\item{lambdax}{A numeric value for the power of the transformation of the explanatory variable (see details).}

\item{shifty}{A numeric shift value for the transformation of the response variable (see details).}

\item{shiftx}{A numeric shift value for the transformation of the explanatory variable (see details).}

\item{show.stats}{A logical indicating if the assumption test p-values should (\code{=TRUE} (default)) be printed on the graphics (see details).}

\item{boxplot}{A logical indicating if the residual plot should be constructed as a boxplot (\code{=TRUE}) or as a traditional residual plot (\code{=FALSE}). Only effective if a one- or two-way ANOVA model is being examined.}

\item{alpha}{A numeric used to decide the significance cutoff when choosing the color to print the assumption test p-values. Only has an effect if \code{show.stats=TRUE}.}

\item{col.hist}{A string used to depict the color of bars in the histogram.}
}
\value{
None. However, a graph, static for \code{assumptionCheck} and dynamic for \code{transChooser}, is produced.
}
\description{
A histogram of residuals and a boxplot of residuals by \dQuote{groups} for ANOVA tests or a residual plot for regression tests is produced. Optionally p-values from the Anderson-Darling test of normality, the outlier test, and Levene's Test for equal variances is shown. The user may also iteratively try power transformations for the response and explanatory variable through arguments in \code{assumptionCheck} or a dynamic graphic using slider bars in \code{transChooser}.
}
\details{
These functions only work for one- and two-way ANOVAs and simple and one- or two-way indicator variable regressions.

Each graphic consists of a histogram of raw residuals on the left and a residual plot (constructed with \code{residPlot} from \pkg{FSA} package) or a boxplot of residuals by group if \code{boxplot=TRUE} in \code{assumptionCheck} or the boxplot check box is selected in the gear box when using \code{transChooser}. P-values from assumption tests will be shown if \code{show.stats=TRUE} in \code{assumptionCheck} or if a check box is selected in the gear box when using \code{transChooser}. The Anderson-Darling p-values is from \code{\link{adTest}}), the outlier test p-value is from \code{\link[car]{outlierTest}} in the \pkg{car} package), and the Levene's Test p-value is from \code{\link[car]{leveneTest}} in the \pkg{car} package).
The \code{lambday} and \code{lambdx} arguments in \code{assumptionCheck} or the slider bar values in the gear box when using \code{transChooser} are values for the power transformation of the response and explanatory variables, respectively. Note that a lambda of 0 corresponds to a natural log transformation. Note that \code{lambdax} is only used if a regression (SLR or IVR) model is being considered.

The \code{shifty} and \code{shiftx} arguments are used to provide a constant value to shift the variable being transformed either left (negative value) or right (positive value) along the respective axis. These values are useful if the original data contains negative numbers as the power transformations generally require non-negative values. Note that \code{shiftx} is only used if a regression (SLR or IVR) model is being considered.
}
\note{
This function is designed to allow \sQuote{newbie} students a method that can be used to quickly test assumptions for simple linear models or to interactively choose appropriate transformations for the response or explanatory variables in these models. This function allows students to choose possible transformations based on an intuitive analysis of diagnostic plots, in contrast, to depending on a non-intuitive method such as the Box-Cox method. While this function can be used for research purposes that was not its intent and that is why it is limited to use with only these four simple models.
}
\examples{
if (require(FSA)) {
data(Mirex)
Mirex$year <- factor(Mirex$year)
Mirex$cyear <- as.character(Mirex$year)

aov1 <- lm(mirex~year,data=Mirex)
assumptionCheck(aov1)
assumptionCheck(aov1,lambday=0)

aov1c <- lm(mirex~cyear,data=Mirex)
assumptionCheck(aov1c)
assumptionCheck(aov1c,lambday=0)

aov2 <- lm(mirex~species*year,data=Mirex)
assumptionCheck(aov2)

slr1 <- lm(mirex~weight,data=Mirex)
assumptionCheck(slr1)
assumptionCheck(slr1,lambday=0)
assumptionCheck(slr1,lambdax=0)

ivr1 <- lm(mirex~weight*year,data=Mirex)
assumptionCheck(ivr1)
}

\dontrun{
# Demonstrates interactive transChooser function
transChooser(aov1)
transChooser(aov2)
transChooser(slr1)
transChooser(ivr1)
}

}
\seealso{
\code{\link[car]{leveneTest}} and \code{\link[car]{outlierTest}}; \code{\link{adTest}}; and \code{\link[MASS]{boxcox}} in \pkg{MASS}.
}
\keyword{dynamic}
\keyword{hplot}
\keyword{models}
