% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compSlopes}
\alias{compSlopes}
\alias{print.compSlopes}
\title{Tests for significant differences among all pairs of slopes in an IVR.}
\usage{
compSlopes(mdl, control = c("fdr", "BH", "BY", "bonferroni", "holm",
  "hochberg", "hommel"), alpha = 0.05, order.slopes = TRUE,
  digits = getOption("digits"))

\method{print}{compSlopes}(x, ...)
}
\arguments{
\item{mdl}{a \code{lm} object.}

\item{control}{a string indicating the method of control to use.  See details.}

\item{alpha}{a decimal numeric indicating the desired level for the
experimentwise error rate}

\item{x}{A \code{compSlopes} object (i.e., returns from \code{compSlopes}).}

\item{order.slopes}{A logical indicating whether the slopes should be order
from smallest to largest upon output.}

\item{digits}{A numeric that controls the number of digits to print.}

\item{\dots}{Other arguments sent to \code{print}.}
}
\value{
A list with three components.  The first component contains the
\code{control} method.  The second component is called \code{comparisons} and
is a data frame that contains the following components:

\tabular{ll}{
\code{comparison} \tab Description of how the difference in levels was computed.\cr
\code{diff} \tab The estimated difference in slope values.\cr
\code{lwr} \tab Lower confidence bound for difference in slope values.\cr
\code{upr} \tab Upper confidence bound for difference in slope values.\cr
\code{raw.p} \tab Unadjusted p-value for testing the difference in slopes is zero.\cr
\code{adj.p} \tab Adjusted p-value for testing the difference in slopes is zero.\cr
}

The third component of the list is called \code{slopes} and is a data frame
that contains the following components:
\tabular{ll}{
\code{level} \tab A level name.\cr
\code{slope} \tab The estimated slope value for the given level.\cr
\code{lwr} \tab Lower confidence bound for difference in slope values.\cr
\code{upr} \tab Upper confidence bound for difference in slope values.\cr
\code{raw.p} \tab Unadjusted p-value for testing the slope is zero.\cr
\code{adj.p} \tab Adjusted p-value for testing the slope is zero.\cr
}

The \code{print} function prints the results nicely.
}
\description{
Tests for significant differences among all pairs of slopes in an indicator
variable regression where the indicator variables all stem from one factor.
}
\details{
In an indicator variable regression the coefficient for the interaction
between the covariate (x) and an indicator variable tests for a difference in
slopes between the level of the indicator variable and the reference level.
Thus, all indicator variables from a particular linear model fit only compare
slopes with the reference level.  Other slope comparisons can be found by
changing the reference level but this requires re-fitting the model.  This
function automates this sequential process and produces a data frame that
shows the estimated difference, an unadjusted confidence interval for the
difference, and the unadjusted p-value for testing that the difference in
slopes is equal to zero.  In addition, the function provides an adjusted
p-value for testing that the difference in slopes is equal to zero and a
statement about whether each null hypothesis is signficant or not.

The adjusted p-values can be computed with a wide variety of methods -- FDR,
BH, BY, bonferroni, holm, hochberg, and hommel.  This function basically
works as a wrapper function that sends the unadjusted \dQuote{raw} p-values
to the \code{p.adjust} function in the base R program.  These functions
should be consulted for further description of the methods used.
}
\note{
This function only works for models with one factor variable.
}
\examples{
require(FSA)      # for fitPlot and Mirex data

data(Mirex)
Mirex$year <- factor(Mirex$year)
# fit an indicator variable regression
lm1 <- lm(mirex~weight*year,data=Mirex)
# visualize the results
fitPlot(lm1)
# compare all pairs of slopes using an FDR control
compSlopes(lm1)
# compare all pairs of slopes using the Holm method of control
compSlopes(lm1,control="holm")
}
\seealso{
\code{fitPlot} in \pkg{FSA}, \code{p.adjust}.
}
\keyword{htest}

