% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\alias{corr.default}
\alias{corr.formula}
\title{Compute correlation coefficients.}
\usage{
corr(x, ...)

\method{corr}{default}(x, y = NULL, use = "everything",
  method = c("pearson", "kendall", "spearman"), digits = NULL, ...)

\method{corr}{formula}(x, data, use = "everything", method = c("pearson",
  "kendall", "spearman"), digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A numeric vector, matrix, or data.frame (see details in \code{\link[stats]{cor}}) or a formula (see details here)}

\item{\dots}{Unused.}

\item{y}{A numeric vector, matrix, or data.frame (see details in \code{\link[stats]{cor}})}

\item{use}{An optional character string that defines a method for computing the correlation in the presence of missing values. This must be one of \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"}, \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}}

\item{method}{A single character string that indicates which correlation coefficient is to be computed. Must be one of \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}}

\item{digits}{A single numeric that indicates the number of decimals to which the result should be rounded.}

\item{data}{An optional data frame that contains the variables in the formula.}
}
\value{
A single numeric if \code{x} and \code{y} are vectors or a numeric matrix if \code{x} and \code{y} are matrices.
}
\description{
Computes the correlation coefficients between \code{y} and \code{x} if these are vectors.  If \code{y} and \code{x} are matrices then the correlation between the columns of \code{y} and the columns of \code{x} are computed.  This functionality is the same as \code{\link[stats]{cor}} in \pkg{stats} except that it adds the formula notation \code{y~x}.
}
\details{
see details in \code{\link[stats]{cor}}.  The formula version is only used with two variables in either the \code{y~x} or \code{~y+x} form.
}
\examples{
## example from stats::cor() ... all return equivalent
df <- data.frame(x=1:10,y=2:11)
cor(df$x,df$y)
corr(df$x,df$y)
corr(y~x,data=df)
corr(~x+y,data=df)
corr(df)

## another example from stats::cor()
cor(longley)
corr(longley)
corr(longley[1:2],longley[3:5])

## handling missing values
set.seed(3473) # to control randomness
df <- data.frame(x=c(runif(9),NA),y=c(NA,runif(9)),z=runif(10),w=runif(10))
corr(~x+y,data=df)
corr(~x+y,data=df,use="complete.obs")
corr(~x+y+z,data=df,use="complete.obs")
corr(~x+y+z,data=df,use="pairwise.complete.obs")
# no messages below (as no missing values)
corr(~w+z,data=df,use="complete.obs")
corr(~w+z,data=df,use="pairwise.complete.obs")

}
\seealso{
See \code{\link[stats]{cor}}.
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}, but this is largely a wrapper to \code{\link[stats]{cor}} from \pkg{stats} that includes a formula version.
}
\keyword{htest}
