% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faqNC.R
\name{faqNC}
\alias{faqNC}
\title{FAQ for using R at Northland College.}
\usage{
faqNC()
}
\value{
None, but a webpage will be opened in the default browswer.
}
\description{
Open web page with FAQ for using R at Northland College.
}
\examples{
\dontrun{
faqNC()
}
}
\keyword{misc}

