% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-teaching.R
\name{iqrCalc}
\alias{iqrCalc}
\alias{print.iqrCalc}
\title{Shows the steps in the manual calculation of the median and IQR.}
\usage{
iqrCalc(x, mdnInBoth = FALSE)

\method{print}{iqrCalc}(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{mdnInBoth}{A logical that indicates whether the median should be placed into both halves or not (DEFAULT) when computing the IQR with odd n.}

\item{\dots}{Other arguments to the generic \code{print} functions (not currently used)}
}
\value{
A list containing the sample size (\code{n}); the sample Q1, median, and Q3 in the name vector \code{vals}; the positions of the sample Q1, median, and Q3 in the named vector \code{pos}; the ordered data (\code{x}); and the ordered lower- (\code{lwr}) and upper-halves (\code{upr}) of the data.
}
\description{
Shows the steps in the manual calculation of the median and IQR.
}
\note{
This function shows the ordered data with the median shown and the ordered lower- and upper-halves of data with the Q1 and Q3 values shown.  This function puts the median into both halves of the data to compute Q1 and Q3.
}
\examples{
## Simple examples
iqrCalc(1:7)
iqrCalc(1:7,mdnInBoth=TRUE)
iqrCalc(1:8)
iqrCalc(1:8,mdnInBoth=TRUE)
iqrCalc(1:9)
iqrCalc(1:9,mdnInBoth=TRUE)
iqrCalc(1:10)
iqrCalc(1:10,mdnInBoth=TRUE)

## Somewhat more realistic
iqrCalc(sample.int(99,11,replace=TRUE))
iqrCalc(sample.int(99,12,replace=TRUE))
iqrCalc(sample.int(99,13,replace=TRUE))
iqrCalc(sample.int(99,14,replace=TRUE))

}
\seealso{
\code{\link{sdCalc}}
}
\keyword{misc}
