% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utils.R
\name{kANOVA}
\alias{kANOVA}
\alias{kGLHT}
\alias{kHtest}
\alias{kREG}
\title{Specific utilities for use in a knitr document.}
\usage{
kANOVA(x)

kGLHT(x, type = c("hypothesis", "confidence"))

kREG(x, digits = max(3, getOption("digits") - 3))

kHtest(x, digits = 4, show.alt = FALSE, ...)
}
\arguments{
\item{x}{An object saved from \code{lm} or \code{glht} or an object of class \code{htest} (e.g., saved from \code{t.test}).}

\item{type}{A string that indicates the type of glht result to extract.}

\item{digits}{Number of decimal places to round the values to.}

\item{show.alt}{A logical that indicates whether the line stating what the alternative hypothesis is should be printed (\code{TRUE}) or not (\code{FALSE}; default).}

\item{\dots}{Additional arguments for the original \code{Stangle} or \code{purl}.}
}
\value{
\itemize{
   \item \code{kANOVA} returns the results of \code{anova} but without the heading attribute. 
   \item \code{kGLHT} returns a matrix of just the hypothesis test or confidence interval results from a \code{glht} object (i.e., all messages that are usually printed are stripped away). 
   \item \code{kREG} returns the results of \code{summary} but without the call and the information about the residuals. 
   \item \code{kHtest} returns the same as \code{stats:::print.htest} except that the name of the test, the data used, and, optionally, the descriptor of the alternative hypothesis are not printed. 
 }
}
\description{
Specific utilities for pretty printing various items in a knitr document.
}
\details{
\itemize{
   \item \code{kANOVA} is used to print \sQuote{pretty} ANOVA tables (i.e., some output removed (see below)).
   \item \code{kGLHT} is used to print \sQuote{pretty} multiple comparison tables (i.e., some output removed (see below)).
   \item \code{kREG} is used to print \sQuote{pretty} summary regression results (i.e., some output removed (see below)).
   \item \code{kHtest} is used to print \sQuote{pretty} hypothesis test (e.g., from \code{\link{t.test}} or \code{\link{chisq.test}}) tables (i.e., some output removed (see below)).
 }
}
\examples{
# None yet

}
\author{
Derek H. Ogle, \email{dogle@northland.edu}
}
\seealso{
See \code{\link[multcomp]{glht}} in \pkg{multcomp} for related functionality.
}
\keyword{hplot}
\keyword{manip}
\keyword{models}

