% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCStats.R
\docType{package}
\name{NCStats}
\alias{NCStats}
\title{Support for statistics classes at Northland College.}
\description{
Support for statistics classes at Northland College.
}
\details{
This package contains functions and simulations to support the Introductory Statistics and Biometry courses at Northland College in Ashland, WI (\url{http://www.northland.edu/}).  \pkg{NCStats} has not been included on CRAN because I have not fully proofed all functions -- they work with the examples that I have developed but have not been thoroughly checked for all possibilities.  Please do not use these functions for research-grade analyses.

Some of the functions in this package are simple modifications of functions from other packages.  I have attempted to give the authors of these other packages appropriate credit while still documenting my modifications.
}
