% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{outlierTest}
\alias{outlierTest}
\title{Performs a test for outliers}
\usage{
outlierTest(...)
}
\arguments{
\item{\dots}{Arguments to pass through to \code{\link[car]{outlierTest}} in \pkg{car}.}
}
\value{
See help for \code{\link[car]{outlierTest}} in \pkg{car}.
}
\description{
Performs a test for outliers
}
\note{
This is a simple pass-through to  \code{\link[car]{outlierTest}} in \pkg{car}.
}
\seealso{
\code{\link[car]{outlierTest}} in \pkg{car}.
}
\keyword{misc}
