% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{percTable}
\alias{percTable}
\title{Express table entries as percentage of marginal table.}
\usage{
percTable(x, margin = NULL, digits = 1, addMargins = !is.na(ncol(x)))
}
\arguments{
\item{x}{A frequency table likely constructed with \code{\link{table}} or \code{\link{xtabs}}.}

\item{margin}{A numeric representing an index, or vector of indices, to generate the margin for -- \code{margin=1} computes row percentages, \code{margin=2} computes column percentages, and \code{margin=NULL} (default) produces table percentages.}

\item{digits}{A numeric indicating the number of decimals to round the percentages to.}

\item{addMargins}{A logical indicating whether marginal totals should be appended to the table or not.  If \code{addMargins=TRUE} then the appended marginal totals will correspond to which margin is chosen with \code{margin=} (as in \code{\link{addMargins}}).}
}
\value{
Same type as \code{x} except with percentages of a margin rather than frequencies.
}
\description{
Same as \code{\link{prop.table}} except that it returns percentages rather than proportions.
}
\examples{
d <- data.frame(Aye=sample(c("Yes","Si","Oui"),177,replace=TRUE),
                Bee=sample(c("Hum","Buzz"),177,replace=TRUE))

## 1-D
( A1 <- table(d$Aye) )
prop.table(A1)
percTable(A1)
percTable(A1,digits=3)
percTable(A1,addMargins=TRUE)
( A2 <- xtabs(~Aye,data=d) )
percTable(A2)

## 2-D
( AB1 <- table(d$Aye,d$Bee) )
percTable(AB1,margin=1)
percTable(AB1,margin=2)
percTable(AB1)
percTable(AB1,digits=3)
percTable(AB1,addMargins=FALSE)

}
\keyword{manip}
