% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{t.test}
\alias{outlierTest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{car}{\code{\link[car]{outlierTest}}}

  \item{mosaic}{\code{\link[mosaic]{t.test}}}
}}

