% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdCalc.R
\name{sdCalc}
\alias{print.sdCalc}
\alias{sdCalc}
\title{Shows the steps in the manual calculation of the standard deviation.}
\usage{
sdCalc(x, digits = getOption("digits"))

print.sdCalc(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{A numeric indicating the number of decimals to round the numeric summaries to.  If left at \code{NULL} (default) then the number of digits will be obtained from \code{getOption('digits')}.}

\item{\dots}{Other arguments to the generic \code{print} functions.  Not currently used.}
}
\value{
A list containing the sample size (\code{n}), sample mean (\code{mean}), data.frame of intermediate calculations (\code{tbl}), and number of digits to print (\code{digits))}.
}
\description{
Shows the steps in the manual calculation of the standard deviation.
}
\note{
This function shows a table of intermediate output in the calculation of the standard deviation.  Used purely to demonstrate the hand-calculation of the standard deviation.  Use \code{sd()} to actually computer the standard deviation.
}
\examples{
## Numeric vector
y <- runif(8)
# typical computation
sd(y)
# this function
sdCalc(y)
# this function, controlling the number of digits
sdCalc(y,digits=4)
}
\seealso{
\code{sd}.
}
\keyword{misc}

