% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-distribs.R
\name{sbeta}
\alias{sbeta}
\alias{sbinom}
\alias{schisq}
\alias{sexp}
\alias{sf}
\alias{sgamma}
\alias{sgeom}
\alias{shyper}
\alias{slnorm}
\alias{snbinom}
\alias{snorm}
\alias{spois}
\alias{st}
\title{Demonstrates the shape of the PDF and CDF for the given distribution.}
\usage{
sbeta()

schisq()

sexp()

sf()

sgamma()

slnorm()

snorm()

st()

sbinom()

sgeom()

shyper()

snbinom()

spois()
}
\value{
None, but a dynamic graphic with slider bars will be produced.
}
\description{
Dynamically demonstrates the effect of parameter choices on the shape of the PDF and CDF for the given distribution.
}
\details{
If the user is using RStudio and the \pkg{manipulate} package is installed then a dynamic graph is produced in the \dQuote{Plots} pane of RStudio.  The plot controls may be accessed through the \dQuote{gear} that is in the upper-left corner of the plot.  The sliders may be used to change the parameters of the distribution.  The real-time updating of the graphic allows the user to determine the effect of changing each parameter on the shape, mean, and standard deviation of the distribution.
}
\examples{
\dontrun{
sbeta()
sbinom()
schisq()
sexp()
sf()
sgamma()
sgeom()
shyper()
slnorm()
snbinom()
snorm()
spois()
st()
}

}
\seealso{
\code{\link[TeachingDemos]{vis.binom}}, \code{\link[TeachingDemos:vis.binom]{vis.gamma}}, \code{\link[TeachingDemos:vis.binom]{vis.normal}}, and \code{\link[TeachingDemos:vis.binom]{vis.t}} of \code{TeachingDemos}.  Also see \code{\link[stats:Beta]{dbeta}}, \code{\link[stats:Binomial]{dbinom}}, \code{\link[stats:Chisquare]{dchisq}}, etc.
}
\keyword{distribution}
\keyword{dynamic}
\keyword{hplot}
