% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_SLR_equation.R
\name{stat_SLR_equation}
\alias{stat_SLR_equation}
\title{Add Regression Line Equation and R-Square to a ggplot.}
\usage{
stat_SLR_equation(
  mapping = NULL,
  data = NULL,
  formula = y ~ x,
  label.x.npc = "left",
  label.y.npc = "top",
  label.x = NULL,
  label.y = NULL,
  geom = "text",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  cf.digits = 2,
  rr.digits = 2,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{formula}{a formula object}

\item{label.x.npc, label.y.npc}{can be \code{numeric} or \code{character} vector of the same length as the number of groups and/or panels. If too short they will be recycled. 
\itemize{ 
  \item If \code{numeric}, value should be between 0 and 1. Coordinates to be used for positioning the label, expressed in "normalized parent coordinates".
  \item If \code{character}, allowed values include: i) one of c('right', 'left', 'center', 'centre', 'middle') for x-axis; ii) and one of c('bottom', 'top', 'center', 'centre', 'middle') for y-axis.}  If too short they will be recycled.}

\item{label.x, label.y}{\code{numeric} Coordinates (in data units) to be used for absolute positioning of the label. If too short they will be recycled.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{cf.digits}{Number of decimals for the regression line coefficients. Can be a vector as long as the coefficients.}

\item{rr.digits}{Number of decimals for R^2 value.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{other arguments to pass to \code{\link[ggplot2]{geom_text}} or \code{\link[ggplot2:geom_text]{geom_label}}.}
}
\description{
Add regression line equation and R^2 to a ggplot. Regression model is fitted using the function \code{\link[stats]{lm}}.
}
\section{Computed variables}{

  \describe{ \item{x}{x position for left edge}
  \item{y}{y position near upper edge}
  \item{eq.label}{equation for the fitted polynomial as a character string to be parsed}
  \item{rr.label}{\eqn{R^2} of the fitted model as a character string to be parsed}
  \item{hjust}{Set to zero to override the default of the "text" geom.}}
}

\examples{
p <- ggplot2::ggplot(data=iris,mapping=ggplot2::aes(x=Sepal.Length,y=Petal.Length)) +
  ggplot2::geom_point() +
  ggplot2::geom_smooth(method="lm") +
  theme_NCStats()
p + stat_SLR_equation()
p + 
  stat_SLR_equation(ggplot2::aes(label=..eq.label..),label.x.npc=0,label.y.npc=1.0) +
  stat_SLR_equation(ggplot2::aes(label=..rr.label..),label.x.npc=0,label.y.npc=0.9)

p2 <- p + ggplot2::facet_wrap(~Species)
p2 + stat_SLR_equation()
p2 + stat_SLR_equation(cf.digits=2)
p2 + stat_SLR_equation(cf.digits=c(3,4))
p2 + 
  stat_SLR_equation(ggplot2::aes(label=..eq.label..),
                    label.x.npc=0,label.y.npc=1.0,cf.digits=c(3,4)) +
  stat_SLR_equation(ggplot2::aes(label=..rr.label..),
                    label.x.npc=0,label.y.npc=0.9,rr.digits=3)

}
\references{
This function is a modification of \code{stat_regline_equation} in \pkg{ggpubr}, but includes \code{cf.digits} and \code{rr.digits}.
}
