% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{theme_NCStats}
\alias{theme_NCStats}
\title{A ggplot2 theme.}
\usage{
theme_NCStats(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/24,
  base_rect_size = base_size/24
)
}
\arguments{
\item{base_size}{The base font size.}

\item{base_family}{The base font family.}

\item{base_line_size}{The base size for line elements.}

\item{base_rect_size}{The base size for rect elements.}
}
\value{
A function that will apply a theme to a ggplot2 object.
}
\description{
A basic ggplot2 theme for use in statistics courses at Northland College.
}
\examples{
p <- ggplot2::ggplot(data=data.frame(x=c(1,2,3,1,3,1),
                                     y=c(1,3,1,2.3,2.3,1)),
                     mapping=ggplot2::aes(x=x,y=y)) +
  ggplot2::geom_polygon()
p
p + theme_NCStats()
}
\keyword{graphics}
\keyword{manip}
