% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transChooser.R
\name{transChooser}
\alias{transChooser}
\title{A dynamic graphic used to select transformations for simple linear models.}
\usage{
transChooser(object, shifty = 0, shiftx = 0, show.stats = TRUE,
  boxplot = TRUE, alpha = 0.05, col.hist = "gray90", ...)
}
\arguments{
\item{object}{An \code{lm} object or formula depicting a one-way or two-way
ANOVA model or a simple linear or indicator variable regression model.}

\item{shifty}{A numeric shift value for the transformation of the response
variable.  See details.}

\item{shiftx}{A numeric shift value for the transformation of the explanatory
variable.  See details.}

\item{show.stats}{A logical indicating if the assumption test p-values should
(\code{=TRUE} (default)) be printed on the graphics.  See details.}

\item{boxplot}{A logical indicating if the residual plot should be constructed as
a boxplot (\code{=TRUE}; default) or as a traditional residual plot (\code{=FALSE}).
Only effective if a one- or two-way ANOVA model is being examined.}

\item{alpha}{A numeric used to decide the significance cutoff when choosing
the color to print the assumption test p-values.  Only has an effect if
\code{show.stats=TRUE}.}

\item{col.hist}{A string used to depict the color of bars in the histogram.}

\item{...}{Other arguments to the generic function.}
}
\value{
None.  However, a dynamic graphic is produced.
}
\description{
A dynamic graphic used to select a response transformation for an ANOVA (one- or
two-way) model or a response or explanatory transformation for a simple linear (SLR)
or indicator variable regression (IVR) model.  The function produces a histogram
of residuals and a residual plot, optionally with assumption checking method p-values,
that are linked to slider bars that allow the user to choose different values of
lambda (the power in a power transformation).  This allows the user to dynamically
select a power transformation that is likely to produce a model that meets the
model assumptions.
}
\details{
This function only works for one- and two-way ANOVAs and simple and indicator
variable regressions.  The indicator variable regression must be either one- or
two-way.  More complicated models can not be considered with transChooser.

Each graphic consists of a histogram of raw residuals on the left and a residual
plot on the right (constructed with \code{residPlot} from the \pkg{FSA} package.)
If \code{show.stats=TRUE} then the p-value from the Anderson-Darling normality
test (as constructed with \code{\link{adTest}}) will be shown above the histogram.
If an ANOVA model is being explored then the p-value from the the outlier test
(as constructed with \code{outliterTest} in the \pkg{car} package) will be shown
above the histogram.  If a regression model is being explored then the outlier
test results will be shown above the residual plot.  If \code{show.stats=TRUE}
and an ANOVA model is being explored then the p-value from a Levene's Test of
equality of variances (as constructed from \code{LeveneTest} in the \pkg{car}
package) will be shown above the residual plot.

If \code{boxplot=TRUE} then the residual plot for the ANOVA models will be a boxplot
of residuals for each group rather than a scatterplot of residuals versus group.
This function does not apply to exploration of regression models.

The dynamics portion of the graphing is accomplished with the \code{slider}
function from either the \pkg{manipulate} package if the user is using the
RStudio IDE or from the \pkg{relax} package if not using the RStudio IDE.
If not in RStudio, then the user must have the \pkg{tcltk} package working
on their machine.  The non-RStudio version will open two windows (the graphic
and a dialog box).  The graphic window may be blank until a slider in the dialog
box has been moved.

The \code{shifty} and \code{shiftx} arguments are used to provide a constant
value to shift the variable being transformed either left (negative value) or
right (positive value) along the respective axis.  These values are useful if
the original data contains negative numbers as the power transformations generally
require non-negative values.  Note that \code{shiftx} is only used if a regression
(SLR or IVR) model is being considered.
}
\note{
This function is designed to allow \sQuote{newbie} students a method that
can be used to interactively choose appropriate transformations for the response
or explanatory variables in one-way ANOVA, two-way ANOVA, simple linear regression,
and indicator variable regressions.  This function allows students to choose
possible transformations based on an intuitive analysis of diagnostic plots,
in contrast, to depending on a non-intuitive method such as the Box-Cox method.
While this function can be used for research purposes that was not its intent
and that is why it is limited to use with only these four simple models.
}
\examples{
require(FSA)   # for Mirex data
data(Mirex)
Mirex$year <- factor(Mirex$year)

if (interactive()) {

## example with one-way ANOVA
lm1 <- lm(mirex~year,data=Mirex)
transChooser(lm1)

## example with two-way ANOVA
lm2 <- lm(mirex~species*year,data=Mirex)
transChooser(lm2)

## example with SLR
lm3 <- lm(mirex~weight,data=Mirex)
transChooser(lm3)

## example with IVR
lm4 <- lm(mirex~weight*year,data=Mirex)
transChooser(lm4)

}
}
\references{
idea came from Fox, J.  1997.  \emph{Applied Regression Analysis,
Linear Models, and Related Methods}.  Sage Publications.  Thousand Oaks, CA.
}
\seealso{
\code{leveneTest}, \code{outlierTest}, and \code{ncvTest} in
\pkg{car}; \code{ad.test} in \pkg{nortest}; and \code{boxcox} in \pkg{MASS}.
}
\keyword{dynamic}
\keyword{hplot}
\keyword{models}

