% Generated by roxygen2 (4.0.0): do not edit by hand
\name{wetPerim}
\alias{plot.wetPerim}
\alias{summary.wetPerim}
\alias{wetPerim}
\title{Computes the wetted perimeter of a stream from depth and distance measurements.}
\usage{
wetPerim(distance, depth)

\method{summary}{wetPerim}(object, detail = TRUE, ...)

\method{plot}{wetPerim}(x, pch = 19, xlab = "Distance (ft)",
  ylab = "Depth (ft)", ...)
}
\arguments{
\item{distance}{A vector of distance measurements at intervals across the cross section.}

\item{depth}{A vector of depth measurements at intervals across the cross-section.}

\item{object}{An object returned from \code{discharge}.}

\item{x}{An object returned from \code{discharge}.}

\item{detail}{A logical indicating whether the details of the discharge calculations should be shown (\code{=TRUE}) or not.}

\item{pch}{The plotting character to be used for the depth profile.  See \code{par}.}

\item{xlab}{A string for labeling the x-axis.}

\item{ylab}{A string for labeling the y-axis.}

\item{\dots}{Other arguments to be passed to the \code{plot} function.}
}
\value{
\code{wetPerim} produces a list with a data frame of the original data, a vector of individual perimeter values, and the total perimeter.  \code{plot} method produce a plot demonstrating the calculations.  \code{summary} returns the total wetted perimeter and detailed intermediate calculations if asked for.
}
\description{
Computes the wetted perimeter of a stream from depth and distance measurements.
}
\examples{
dist1 <- c(seq(0,22,2),24.7)
dep1 <- c(0,.9,1.1,1.1,1.1,.9,.9,.9,.9,.9,.9,.7,0)
vel1 <- c(0,1.01,1.07,1.22,1.22,1.36,1.28,1.38,1.29,0.88,0.52,0,0)
ex1w <- wetPerim(dist1,dep1)
summary(ex1w)
plot(ex1w)

dist2 <- c(seq(0,18,2),19.7)
dep2 <- c(0,1.2,1.5,1.4,1.4,1.4,1.5,1.7,2,1.5,0)
vel2 <- c(0,.53,.76,.95,.90,.98,1.01,.9,.9,0,.0)
ex2w <- wetPerim(dist2,dep2)
summary(ex2w)
plot(ex2w)
}
\keyword{hplot}
\keyword{models}

