% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{z.test}
\alias{z.test}
\title{Z test for known population standard deviation}
\arguments{
\item{x}{Vector of data values}

\item{mu}{Hypothesized mean of the population}

\item{sd}{Known standard deviation of the population}

\item{alternative}{Direction of the alternative hypothesis}

\item{conf.level}{Confidence level for the interval computation}

\item{\dots}{Additional arguments are silently ignored}
}
\value{
An object of class \code{htest} containing the results
}
\description{
Compute the test of hypothesis and compute confidence interval on the mean of a population when the standard deviation of the population is known.
}
\details{
Most introductory statistical texts introduce inference by using the Z test and Z based confidence intervals based on knowing the population standard deviation. Most statistical packages do not include functions to do Z tests because the T test is usually more appropriate for real world situations. This function is meant to be used during that short period of learning when the student is learning about inference using Z procedures, but has not learned the T based procedures yet. Once the student has learned about the T distribution the \code{\link[stats]{t.test}} function should be used instead of this one (but the syntax is very similar, so this function should be an appropriate introductory step to learning \code{\link[stats]{t.test}}).
}
\note{
This function should be used for learning only, real data should generally use \code{\link[stats]{t.test}}.
}
\examples{
z.test(rnorm(25,100,5),99,5)

}
\seealso{
\code{\link{t.test}}
}
\author{
This is \code{\link[TeachingDemos]{z.test}} in \pkg{TeachingDemos} but imported and exported here.
}
\keyword{htest}
