% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsPostHoc-utils.R
\name{glhtSig}
\alias{glhtSig}
\alias{glhtSig.glht}
\alias{glhtSig.glht}
\title{Extracts names for significantly different comparisons.}
\usage{
glhtSig(object, ...)

\method{glhtSig}{glht}(object, type = c("single-step", "Shaffer", "Westfall",
  "free", stats::p.adjust.methods), alpha = 0.05, ...)
}
\arguments{
\item{object}{An object saved from \code{\link[multcomp]{glht}} from the \pkg{multcomp} package.}

\item{\dots}{Other arguments to pass through to \code{\link[multcomp]{summary.glht}}.}

\item{type}{A function for computing p-values (see \code{\link[multcomp]{glht}}).}

\item{alpha}{A numeric indicated the rejection criterion to be used for identifying significant comparisons. Defaults to \code{0.05}.}
}
\value{
A vector containing the names of the paired comparisons that have a p-value less than \code{alpha}.
}
\description{
Extracts names for significantly different comparisons from a \code{glht} object.
}
\note{
This can be very slow to process. Be patient.
}
\examples{
## NONE YET

}
\seealso{
\code{\link[multcomp]{glht}} and \code{\link[multcomp]{summary.glht}} in \pkg{multcomp}.
}
\keyword{misc}
