% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify.formula.R
\name{identify.formula}
\alias{identify.formula}
\title{Identify points in a plot using a formula.}
\usage{
\method{identify}{formula}(x, data = NULL, ...)
}
\arguments{
\item{x}{A formula of the form \code{y~x}.}

\item{data}{The data frame from which the formula should be evaluated.}

\item{\dots}{Other arguments to be passed to \code{\link[graphics]{identify}}.}
}
\value{
See \code{identify}
}
\description{
Reads the position of the graphics pointer when the (first) mouse button is pressed, searches the coordinates given in x and y for the point closest to the pointer, and return the index if this point is close enough to the pointer.
}
\details{
This function is meant to make it easier to call \code{\link[graphics]{identify}} after \code{plot} has been called using a formula and the \code{data=} argument.

A two dimensional plot must be active and the vectors in \code{x} and data frame  in \code{data} must correspond to the x- and y-axes and the data of the plot.
}
\examples{
## Not run by examples().  Copy and try in an interactive R session
\dontrun{
df <- data.frame(x=runif(10),y=runif(10),grp=factor(rep(c("Yes","No"),each=5)))
plot(y~x,data=df)
identify(y~x,data=df)
}
}
\seealso{
\code{identify}
}
\keyword{hplot}

