% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/levenesTest.R
\name{levenesTest}
\alias{levenesTest}
\title{Perform Levene's test for homogeneity of variances}
\usage{
levenesTest(...)
}
\arguments{
\item{\dots}{Arguments to pass through to \code{leveneTest} in \pkg{car}.}
}
\value{
See help for \code{leveneTest} in \pkg{car}.
}
\description{
Perform Levene's test for homogeneity of variances
}
\note{
This is a simple pass-through to  \code{leveneTest} in \pkg{car} for those
of us that always put the \sQuote{s} on \sQuote{Levenes}.
}
\seealso{
\code{leveneTest} in \pkg{car}.
}
\keyword{misc}

