% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SquareLakePopn.R
\docType{data}
\name{SquareLakePopn}
\alias{SquareLakePopn}
\title{Biological characteristics of the "Square Lake" population.}
\format{A data frame of 1014 observations on the following 6 variables:
  \describe{
    \item{tl}{Total length (mm) of an individual fish.}
    \item{fl}{Fork length (mm) of an individual fish.}
    \item{sl}{Standard length (mm) of an individual fish.}
    \item{w}{Weight (g) of an individual fish.}
    \item{age}{Age of an individual fish.}
    \item{sex}{Sex of an individual fish (\code{F}=Female and \code{M}=Male).}
  }}
\description{
Simulated data for the population of 1014 fish in \dQuote{Square Lake}.
}
\examples{
data(SquareLakePopn)
str(SquareLakePopn)
head(SquareLakePopn)

}
\keyword{datasets}
