% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ztest.R
\name{z.test}
\alias{z.test}
\title{Z test for known population standard deviation}
\usage{
z.test(x, mu = 0, sd, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95, ...)
}
\arguments{
\item{x}{Vector of data values.}

\item{mu}{Hypothesized mean of the population.}

\item{sd}{Known standard deviation of the population.}

\item{alternative}{Direction of the alternative hypothesis.}

\item{conf.level}{Confidence level for the interval computation.}

\item{\dots}{Additional arguments are silently ignored.}
}
\value{
An object of class \code{htest} containing the results
}
\description{
Compute the test of hypothesis and compute confidence interval on the mean of
a population when the standard deviation of the population is known.
}
\details{
Most introductory statistical texts introduce inference by using the Z test
and Z based confidence intervals based on knowing the population standard
deviation.  Most statistical packages do not include functions to do Z tests
because the T test is usually more appropriate for real world situations.
This function is meant to be used during that short period of learning when
the student is learning about inference using Z procedures, but has not
learned the T based procedures yet.  Once the student has learned about the T
distribution the \code{t.test()} function should be used instead of this one
(but the syntax is very similar, so this function should be an appropriate
introductory step to learning \code{t.test()}).
}
\note{
This function should be used for learning only, real data should
generally use \code{t.test()}.
}
\examples{
x <- rnorm(25, 100, 5)
z.test(x, 99, 5)
}
\author{
Greg Snow \email{greg.snow@imail.org} with a slight modification (removed
the stdev argument) by the package author.
}
\seealso{
\code{\link{t.test}}
}
\keyword{htest}

